add vcredist install and silence output
Problem * If using Write-Output the following error occurs: "Error on The OS handle's position is not what FileStream expected. Do not use a handle simultaneously in one FileStream and in Win32 code or another FileStream. This may cause data loss". So this is commented out till a PowerShell Guru can tell me how to prevent this.